# bpo-46848: Optimize mmap.find() with memmem(3)