--- a/configure
+++ b/configure
@@ -14118,6 +14118,42 @@ fi
 
 
 
+  { $as_echo "$as_me:${as_lineno-$LINENO}: checking for memmem" >&5
+$as_echo_n "checking for memmem... " >&6; }
+if ${ac_cv_func_memmem+:} false; then :
+  $as_echo_n "(cached) " >&6
+else
+  cat confdefs.h - <<_ACEOF >conftest.$ac_ext
+/* end confdefs.h.  */
+#include <string.h>
+int
+main ()
+{
+void *x=memmem
+  ;
+  return 0;
+}
+_ACEOF
+if ac_fn_c_try_compile "$LINENO"; then :
+  ac_cv_func_memmem=yes
+else
+  ac_cv_func_memmem=no
+fi
+rm -f core conftest.err conftest.$ac_objext conftest.$ac_ext
+
+fi
+{ $as_echo "$as_me:${as_lineno-$LINENO}: result: $ac_cv_func_memmem" >&5
+$as_echo "$ac_cv_func_memmem" >&6; }
+  if test "x$ac_cv_func_memmem" = xyes; then :
+
+$as_echo "#define HAVE_MEMMEM 1" >>confdefs.h
+
+fi
+
+
+
+
+
   { $as_echo "$as_me:${as_lineno-$LINENO}: checking for kqueue" >&5
 $as_echo_n "checking for kqueue... " >&6; }
 if ${ac_cv_func_kqueue+:} false; then :