This merge requests optimizes `mmap.find()` by using libc's `memmem(3)` function (which is optimized) instead of using a trivial for loop. I could not find an equivalent solution for `mmap.rfind()` yet. 

The following snippet runs 100% faster with this patch applied. The file used is a logfile of 2.3 GB size

``` python
with open(sys.argv[1], 'r+b') as file_:
    with mmap.mmap(file_.fileno(), 0, access=mmap.ACCESS_READ) as file:
        cur_offset = 0
        while True:
            cur_offset = file.find(b"\n", cur_offset)
            if cur_offset == -1:
                break
            cur_offset += 1
```


With this patch applied:
```
rumpelsepp@kronos ~/P/v/c/debug (main)> time ./python test.py /tmp/log

________________________________________________________
Executed in    3,12 secs    fish           external
   usr time    2,94 secs    1,34 millis    2,94 secs
   sys time    0,15 secs    0,00 millis    0,15 secs

```

Without this patch:
```
rumpelsepp@kronos ~/P/v/c/debug (mmap)> time ./python test.py /tmp/log

________________________________________________________
Executed in    6,47 secs    fish           external
   usr time    6,53 secs  248,91 millis    6,29 secs
   sys time    0,18 secs   32,26 millis    0,15 secs
```

<!-- issue-number: [bpo-46848](https://bugs.python.org/issue46848) -->
https://bugs.python.org/issue46848
<!-- /issue-number -->