@sweeneyde 2022-02-24
Another thing you could try out and compare is using the fastsearch functions defined in Objects/stringlib, which should generally be faster than that naive loop and also platform independent.