@rumpelsepp 2022-03-01
https://github.com/python/cpython/pull/31625 provides a platform independent proposal.